move available fonts to redux store
