+++ baby-gru/src/store/labelSettingsSlice.ts
@@ -18,6 +18,9 @@ export const labelSettingsSlice = createSlice({
     emptyAvailableFonts: (state) => {
       return {...state, availableFonts: [ ]}
     },
+    addAvailableFontList: (state, action: {payload: string[], type: string}) => {
+      return {...state, availableFonts: [...state.availableFonts, ...action.payload]}
+    },
     setAtomLabelDepthMode: (state, action: {payload: boolean, type: string}) => {
       return {...state, atomLabelDepthMode: action.payload}
     },
@@ -30,6 +33,6 @@ export const labelSettingsSlice = createSlice({
   }
 })
 
-export const { setAtomLabelDepthMode, setGLLabelsFontFamily, setGLLabelsFontSize } = labelSettingsSlice.actions
+export const { addAvailableFontList, setAtomLabelDepthMode, setGLLabelsFontFamily, setGLLabelsFontSize } = labelSettingsSlice.actions
 
 export default labelSettingsSlice.reducer
